Persistent Dao realisation
