+++ src/main/java/ru/vk/itmo/kovalevigor/UtilsMemorySegment.java
@@ -1,0 +1,23 @@
+package ru.vk.itmo.kovalevigor;
+
+import java.lang.foreign.MemorySegment;
+import java.lang.foreign.ValueLayout;
+
+public class UtilsMemorySegment {
+
+    private static byte getByte(final MemorySegment memorySegment, final long offset) {
+        return memorySegment.get(ValueLayout.JAVA_BYTE, offset);
+    }
+    public static int compare(final MemorySegment lhs, final MemorySegment rhs) {
+        final long mismatch = lhs.mismatch(rhs);
+        final long lhsSize = lhs.byteSize();
+        final long rhsSize = rhs.byteSize();
+        final long minSize = Math.min(lhsSize, rhsSize);
+        if (mismatch == -1) {
+            return 0;
+        } else if (minSize == mismatch) {
+            return Long.compare(lhsSize, rhsSize);
+        }
+        return Byte.compare(getByte(lhs, mismatch), getByte(rhs, mismatch));
+    }
+}
