+++ src/main/java/ru/vk/itmo/test/kovalevigor/DaoFactoryImpl.java
@@ -4,7 +4,6 @@ import ru.vk.itmo.Config;
 import ru.vk.itmo.Dao;
 import ru.vk.itmo.Entry;
 import ru.vk.itmo.kovalevigor.DaoImpl;
-import ru.vk.itmo.kovalevigor.PersistentDao;
 import ru.vk.itmo.test.DaoFactory;
 
 import java.io.IOException;
@@ -20,12 +19,12 @@ public class DaoFactoryImpl implements DaoFactory.Factory<MemorySegment, Entry<M
 
     @Override
     public Dao<MemorySegment, Entry<MemorySegment>> createDao() {
-        return new DaoImpl();
+        return null;
     }
 
     @Override
     public Dao<MemorySegment, Entry<MemorySegment>> createDao(Config config) throws IOException {
-        return new PersistentDao(config);
+        return new DaoImpl(config);
     }
 
     @Override
